agent_platform/requirements.txt
# Titan Platform - Agent Platform Dependencies

# Google ADK
google-adk>=0.1.0
google-generativeai>=0.8.0

# Market Data
yfinance>=0.2.50
pandas>=2.2.0
numpy>=1.26.0

# Technical Analysis
pandas-ta>=0.3.14b

# Vector Database (for Memory Bank)
chromadb>=0.5.0

# Search & Scraping
duckduckgo-search>=6.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# API & Web
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Utilities
python-dateutil>=2.8.0
